requirements.txt
streamlit
sentence-transformers
numpy
transformers
torch
PyPDF2
openai==0.28
python-dotenv
pinecone
